[#2][#3] make a simple line chart
- changed the area from rectangle to actual area with start and end
  point
- used actual class Line for LineChart
- let Lines paint directly to a coordinate system

FIXME: SWT cannot draw doubles, only ints (so it's useless for sin())

+++ swt-charts/src/main/java/de/slothsoft/charts/swt/SwtGraphicContext.java
@@ -78,7 +78,7 @@ public class SwtGraphicContext implements GraphicContext {
 
 	@Override
 	public void scale(double x, double y) {
-		doToTransform(t -> t.scale((int) x, (int) y));
+		doToTransform(t -> t.scale((float) x, (float) y));
 	}
 
 	@Override
